pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "pySC/__init__.py"

[tool.hatch.build.targets.sdist]
exclude = [
  "/.github",
  "/doc",
  "/tests",
]

[tool.hatch.build.targets.wheel]
packages = ["pySC"]

[project]
name = "accelerator-commissioning"
description = "A Python Simulated Commissioning toolkit for synchrotrons (inspired by https://github.com/ThorstenHellert/SC)"
readme = "README.md"
requires-python = ">=3.9"
dynamic = ["version"]
#license = {file = "LICENSE.txt"}
keywords = ["Accelerator", "Synchrotron", "Commissioning", ]
maintainers = [
    {name = "Konstantinos Paraschou", email = "konstantinos.paraschou@desy.de"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "numpy>=2.0.0",
    "scipy>=1.11.4",
    "matplotlib>=3.7.3",
    "accelerator-toolbox>=0.6.1",
    "pydantic",
    "PyYAML>=6.0.2",
    "rich",
    "h5py"
]

[project.optional-dependencies]
test = [
    "pytest>=7.4",
    "pytest-cov>=3.0",
]
doc = [
    "sphinx",
    "travis-sphinx",
    "sphinx-rtd-theme"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: tests that involve AT tracking (deselect with '-m \"not slow\"')",
    "regression: regression tests for known bugs",
]
addopts = "--tb=short -q"

[project.urls]
Homepage = "https://github.com/kparasch/pySC"
Repository = "https://github.com/kparasch/pySC.git"
